fix: Only create links durin package installs if they don't exist

+++ packages/pkg-postinst-core.sh
@@ -14,10 +14,10 @@ if ! getent passwd opennms >/dev/null 2>&1; then
           opennms
 fi
 
-ln -s /var/log/opennms /opt/opennms/logs
-ln -s /var/lib/opennms/rrd /opt/opennms/share/rrd
-ln -s /var/lib/opennms/reports /opt/opennms/share/reports
-ln -s /var/lib/opennms/deploy /opt/opennms/deploy
+[ ! -e /opt/opennms/logs ] && ln -s /var/log/opennms /opt/opennms/logs
+[ ! -e /opt/opennms/share/rrd ] && ln -s /var/lib/opennms/rrd /opt/opennms/share/rrd
+[ ! -e /opt/opennms/share/reports ] && ln -s /var/lib/opennms/reports /opt/opennms/share/reports
+[ ! -e /opt/opennms/deploy ] && ln -s /var/lib/opennms/deploy /opt/opennms/deploy
 chown -R opennms:opennms /opt/opennms
 chown -R opennms:opennms /var/log/opennms
 chown -R opennms:opennms /var/lib/opennms
