+++ packages/pkg-postinst-minion.sh
@@ -14,8 +14,8 @@ if ! getent passwd minion >/dev/null 2>&1; then
           minion
 fi
 
-ln -s /var/lib/minion/deploy /opt/minion/deploy
-ln -s /var/log/minion /opt/minion/data/log
+[ ! -e /opt/minion/deploy ] && ln -s /var/lib/minion/deploy /opt/minion/deploy
+[ ! -e /opt/minion/data/log ] && ln -s /var/log/minion /opt/minion/data/log
 chown -R minion:minion /opt/minion
 chown -R minion:minion /var/lib/minion
 chown -R minion:minion /var/log/minion
